moving to vectors

+++ src/objects/sphere.c
@@ -36,7 +36,7 @@ char		*ft_parse_sphere(char *attr, t_scene *scn)
 	attr = ft_get_curve(attr, '{');
 	ft_get_attr_in_scope(attr, "origin:", (void *)(&(sph->origin)), PNT);
 	ft_get_attr_in_scope(attr, "radius:", (void *)(&(sph->radius)), FLT);
-	sph->radius = ft_limitf(0.0, DBL_MAX, sph->radius);
+	sph->radius = (float)ft_limitf(0.0f, FLT_MAX, sph->radius);
 	sph->origin = ft_3_vector_add(sph->origin, obj->translate);
 	obj->fig = sph;
 	ft_lstpush(&(scn->objs), ft_nodenew((void *)obj, sizeof(obj)));
